change ui.js and bad_word_list.js like block_other_players one
